split tag handling
